Fix DI for sponsor commands, refactor a bit
The existing commands were not resolving anymore since we removed the ICommandApp from the DI container. This is because the lifetime of services gets all screwed up if you use Run() on the app more than once. So we instead execute the relevant commands directly.
